Moved actions to corresponding controllers
